requirements.txt
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
python-dateutil==2.8.2
sgmllib3k==1.0.0
Pillow>=10.0.0
